Using when and cleanup
