+++ src/NServiceBus.Transport.SQS.AcceptanceTests/Sending/When_using_large_message_with_serverside_aes.cs
@@ -1,20 +1,18 @@
 namespace NServiceBus.AcceptanceTests
 {
-    using System;
-    using System.Security.Cryptography;
     using System.Threading.Tasks;
     using AcceptanceTesting;
     using Amazon.S3;
-    using Amazon.S3.Model;
     using EndpointTemplates;
     using NUnit.Framework;
 
-    public class Sending_large_message_using_customer_provided_aes : NServiceBusAcceptanceTest
+    public class When_using_large_message_with_serverside_aes : NServiceBusAcceptanceTest
     {
         [Test]
         public async Task Should_receive_message()
         {
             var payloadToSend = new byte[PayloadSize];
+
             var context = await Scenario.Define<Context>()
                 .WithEndpoint<Endpoint>(b => b.When(session => session.SendLocal(new MyMessageWithLargePayload
                 {
@@ -27,22 +25,15 @@ namespace NServiceBus.AcceptanceTests
             Assert.AreEqual(payloadToSend, context.ReceivedPayload, "The large payload should be handled correctly using the kms encrypted S3 bucket");
 
             var s3Client = ConfigureEndpointSqsTransport.CreateS3Client();
-            var getObjectResponse = await s3Client.GetObjectAsync(new GetObjectRequest
-            {
-                BucketName = BucketName,
-                Key = $"{ConfigureEndpointSqsTransport.S3Prefix}/{context.MessageId}",
-                ServerSideEncryptionCustomerMethod = ServerSideEncryptionCustomerMethod.AES256,
-                ServerSideEncryptionCustomerProvidedKey = Base64Key,
-            });
+            var getObjectResponse = await s3Client.GetObjectAsync(BucketName, $"{ConfigureEndpointSqsTransport.S3Prefix}/{context.MessageId}");
 
-            Assert.AreEqual(ServerSideEncryptionCustomerMethod.AES256, getObjectResponse.ServerSideEncryptionCustomerMethod);
-            Assert.IsNull(getObjectResponse.ServerSideEncryptionMethod);
+            Assert.AreEqual(ServerSideEncryptionMethod.AES256, getObjectResponse.ServerSideEncryptionMethod);
+            Assert.AreEqual(ServerSideEncryptionCustomerMethod.None, getObjectResponse.ServerSideEncryptionCustomerMethod);
         }
 
         const int PayloadSize = 150 * 1024;
 
         static string BucketName;
-        static string Base64Key;
 
         public class Context : ScenarioContext
         {
@@ -52,42 +43,32 @@ namespace NServiceBus.AcceptanceTests
 
         public class Endpoint : EndpointConfigurationBuilder
         {
-            public Endpoint()
-            {
+            public Endpoint() =>
                 EndpointSetup<DefaultServer>(c =>
                 {
                     var transportConfig = c.ConfigureSqsTransport();
 
                     BucketName = $"{ConfigureEndpointSqsTransport.S3BucketName}";
 
-                    var aesEncryption = Aes.Create();
-                    aesEncryption.KeySize = 256;
-                    aesEncryption.GenerateKey();
-                    Base64Key = Convert.ToBase64String(aesEncryption.Key);
-
                     transportConfig.S3 = new S3Settings(BucketName, ConfigureEndpointSqsTransport.S3Prefix, ConfigureEndpointSqsTransport.CreateS3Client())
                     {
-                        Encryption = new S3EncryptionWithCustomerProvidedKey(ServerSideEncryptionCustomerMethod.AES256, Base64Key)
+                        Encryption = new S3EncryptionWithManagedKey(ServerSideEncryptionMethod.AES256)
                     };
                 });
-            }
 
             public class MyMessageHandler : IHandleMessages<MyMessageWithLargePayload>
             {
-                Context testContext;
-
-                public MyMessageHandler(Context testContext)
-                {
-                    this.testContext = testContext;
-                }
+                public MyMessageHandler(Context testContext) => this.testContext = testContext;
 
                 public Task Handle(MyMessageWithLargePayload messageWithLargePayload, IMessageHandlerContext context)
                 {
                     testContext.MessageId = context.MessageId;
                     testContext.ReceivedPayload = messageWithLargePayload.Payload;
 
-                    return Task.FromResult(0);
+                    return Task.CompletedTask;
                 }
+
+                readonly Context testContext;
             }
         }
