+++ src/NServiceBus.Transport.SQS.AcceptanceTests/Sending/When_using_large_message_with_unencrypted_bucket.cs
@@ -1,17 +1,17 @@
-namespace NServiceBus.AcceptanceTests
+﻿namespace NServiceBus.AcceptanceTests
 {
     using System.Threading.Tasks;
     using AcceptanceTesting;
-    using Amazon.S3;
     using EndpointTemplates;
     using NUnit.Framework;
 
-    public class Sending_large_message_using_serverside_aes : NServiceBusAcceptanceTest
+    public class When_using_large_message_with_unencrypted_bucket : NServiceBusAcceptanceTest
     {
         [Test]
         public async Task Should_receive_message()
         {
             var payloadToSend = new byte[PayloadSize];
+
             var context = await Scenario.Define<Context>()
                 .WithEndpoint<Endpoint>(b => b.When(session => session.SendLocal(new MyMessageWithLargePayload
                 {
@@ -21,18 +21,14 @@ namespace NServiceBus.AcceptanceTests
                 .Done(c => c.ReceivedPayload != null)
                 .Run();
 
-            Assert.AreEqual(payloadToSend, context.ReceivedPayload, "The large payload should be handled correctly using the kms encrypted S3 bucket");
+            Assert.AreEqual(payloadToSend, context.ReceivedPayload, "The large payload should be handled correctly using the unencrypted S3 bucket");
 
             var s3Client = ConfigureEndpointSqsTransport.CreateS3Client();
-            var getObjectResponse = await s3Client.GetObjectAsync(BucketName, $"{ConfigureEndpointSqsTransport.S3Prefix}/{context.MessageId}");
 
-            Assert.AreEqual(ServerSideEncryptionMethod.AES256, getObjectResponse.ServerSideEncryptionMethod);
-            Assert.AreEqual(ServerSideEncryptionCustomerMethod.None, getObjectResponse.ServerSideEncryptionCustomerMethod);
+            Assert.DoesNotThrowAsync(async () => await s3Client.GetObjectAsync(ConfigureEndpointSqsTransport.S3BucketName, $"{ConfigureEndpointSqsTransport.S3Prefix}/{context.MessageId}"));
         }
 
         const int PayloadSize = 150 * 1024;
-
-        static string BucketName;
 
         public class Context : ScenarioContext
         {
@@ -42,37 +38,26 @@ namespace NServiceBus.AcceptanceTests
 
         public class Endpoint : EndpointConfigurationBuilder
         {
-            public Endpoint()
-            {
+            public Endpoint() =>
                 EndpointSetup<DefaultServer>(c =>
                 {
-                    var transportConfig = c.ConfigureSqsTransport();
-
-                    BucketName = $"{ConfigureEndpointSqsTransport.S3BucketName}";
-
-                    transportConfig.S3 = new S3Settings(BucketName, ConfigureEndpointSqsTransport.S3Prefix, ConfigureEndpointSqsTransport.CreateS3Client())
-                    {
-                        Encryption = new S3EncryptionWithManagedKey(ServerSideEncryptionMethod.AES256)
-                    };
+                    c.ConfigureSqsTransport().S3
+                        = new S3Settings(ConfigureEndpointSqsTransport.S3BucketName, ConfigureEndpointSqsTransport.S3Prefix, ConfigureEndpointSqsTransport.CreateS3Client());
                 });
-            }
 
             public class MyMessageHandler : IHandleMessages<MyMessageWithLargePayload>
             {
-                Context testContext;
-
-                public MyMessageHandler(Context testContext)
-                {
-                    this.testContext = testContext;
-                }
+                public MyMessageHandler(Context testContext) => this.testContext = testContext;
 
                 public Task Handle(MyMessageWithLargePayload messageWithLargePayload, IMessageHandlerContext context)
                 {
                     testContext.MessageId = context.MessageId;
                     testContext.ReceivedPayload = messageWithLargePayload.Payload;
 
-                    return Task.FromResult(0);
+                    return Task.CompletedTask;
                 }
+
+                readonly Context testContext;
             }
         }
